batch rename the second
