+++ .gitattributes
@@ -1,2 +1,2 @@
-pymip/_version.py export-subst
+xmip/_version.py export-subst
 * text=auto eol=lf
