+++ MANIFEST.in
@@ -1,1 +1,1 @@
-recursive-include pymip *
+recursive-include xmip *
