+++ ci/environment-upstream-dev.yml
@@ -1,4 +1,4 @@
-name: test_env_pymip
+name: test_env_xmip
 channels:
   - conda-forge
 dependencies:
@@ -8,7 +8,7 @@ dependencies:
   - pip
   - cartopy #installing this without conda is a nightmare, so ill leave it here
   - xesmf # same here
-  - rasterio # Trying to get around an apparent bug with py 3.10 + pip + rasterio (https://github.com/jbusecke/pymip/pull/231#issuecomment-1132190649)
+  - rasterio # Trying to get around an apparent bug with py 3.10 + pip + rasterio (https://github.com/jbusecke/xmip/pull/231#issuecomment-1132190649)
   # These are only a patch until xesmf > 0.6.2 is released. See https://github.com/pangeo-data/xESMF/issues/170
   - numba
   - sparse>=0.8.0
