+++ docs/environment.yml
@@ -1,4 +1,4 @@
-name: pymip_docs
+name: xmip_docs
 channels:
   - conda-forge
 dependencies:
