+++ setup.cfg
@@ -18,14 +18,14 @@ ignore =
 max-line-length = 105
 select = C,E,F,W,B,B950
 ignore = E203, E501, W503
-exclude = pymip/_version.py
+exclude = xmip/_version.py
 
 
 [metadata]
-name = pymip
+name = xmip
 description = Analysis ready CMIP6 data the easy way
-author = pymip developers
-url=https://github.com/jbusecke/pymip
+author = xmip developers
+url=https://github.com/jbusecke/xmip
 long_description = file: README.md
 long_description_content_type = text/markdown
 license = MIT
