+++ tests/test_preprocessing.py
@@ -4,8 +4,8 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from pymip.postprocessing import exact_attrs
-from pymip.preprocessing import (
+from xmip.postprocessing import exact_attrs
+from xmip.preprocessing import (
     broadcast_lonlat,
     cmip6_renaming_dict,
     combined_preprocessing,
