+++ tests/test_preprocessing_cloud.py
@@ -7,9 +7,9 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from pymip.grids import combine_staggered_grid
-from pymip.preprocessing import _desired_units, _drop_coords, combined_preprocessing
-from pymip.utils import google_cmip_col, model_id_match
+from xmip.grids import combine_staggered_grid
+from xmip.preprocessing import _desired_units, _drop_coords, combined_preprocessing
+from xmip.utils import google_cmip_col, model_id_match
 
 
 pytest.importorskip("gcsfs")
@@ -53,7 +53,7 @@ def data(
         else:
             ##### debugging options
             # @charlesbluca suggested this to make this work in GHA
-            # https://github.com/jbusecke/pymip/pull/62#issuecomment-741928365
+            # https://github.com/jbusecke/xmip/pull/62#issuecomment-741928365
             mm = fsspec.get_mapper(
                 cat.df["zstore"][0]
             )  # think you can pass in storage options here as well?
