+++ xmip/specs/staggered_grid_config.yaml
